requirements.txt
python-dotenv~=0.16.0
pyTelegramBotAPI=3.7.7